Divide the file nokogiri.rb into three files: document.rb, header.rb and link.rb for each class separatetly. Also add a check for existence of input file
